+++ lib/omniauth-bookingsync.rb
@@ -1,2 +1,3 @@
 require "omniauth/bookingsync/version"
+require "omniauth/bookingsync/identifier"
 require "omniauth/strategies/bookingsync"