+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 #master
 
+* Add support for omniauth-oauth2 1.4
+* Update gems dependencies
+* Add Ruby 2.4 to Travis
+* Add account auto-selection during the authorization process
+
+# 0.4.0
+
+* Update Readme
 * bump oauth2 to 1.2
 * add Ruby 2.3.1 to travis, drop Ruby 2.0 and 2.1 (end of maintenance)
 * Update example application to work over SSL